Release version 2.10.0+2
Upgrade to upstream's commit fa83b0fb5e91 (type_info: remove non-UTF-8 characters)

The reason for this upgrade is to apply the bug-fix for the issue #1
(Non-UTF-8 characters in type_info.hpp). Note that this way we fast forward
through the upsteam's 2.11.0 release:

2.10.0+1 -> 432c55de9b9a (build system: update gtest upstream branch name)
            86fb93d0ef03, v2.11.0_libuv_v1.43 (update to libuv v1.43)
2.10.0+2 -> fa83b0fb5e91 (type_info: remove non-UTF-8 characters)

Note that 86fb93d0ef03 just increments the version to 2.11.0 and switches to
version 1.43 of the libuv dependency in the cmake/conan files.

Also note that we don't increment our own version to 2.11.0 just yet to make
sure that 2.10.0 is also fixed. We will release 2.11.0 while bumping libuv
dependency version constraint to ~1.43.0 (which is not packaged yet).

Revert commit f69df47241e8 (Add MSVC utf-8 flag).

Add tests/.gitignore file.